+++ functions/src/private/content.ts
@@ -110,185 +110,6 @@ export const getListOfSubjectSecretPrompt = async (
   };
 };
 
-export const getBodyPrompt = async (
-  subtitle: string,
-  mission: string,
-  context: string,
-  targetAudience: string,
-  listSubtitle: string[],
-): Promise<{ system: string; user: string }> => {
-  return {
-    system: `Create a text, this text is a section of an article.
-    It should respect the requirement that follow:
-    "
-    - This section should have a minimum of 4 paragraphs.
-    - Each paragraph will be short and concise, to be easy to read.
-    - You will be given a subtitle of the article and the mission of the article, the text must reflect of those.
-    - You will also be given a context, this context will act as a set of instructions for the text you will create.
-    - Make sure the content is related to the mission of the article: "${mission}".
-    - Your task is to generate engaging and informative content for the subtitle.
-    - The content should be designed to hold the reader’s attention and provide valuable information without overwhelming them.
-    - The text output is destined to bring organic traffic to a website.
-    - It should have a lot of different keywords related to the context.
-    - The people that will read the text don't necessarily have a lot of knowledge on the subject.
-    - You should grab the readers attention with genuine interest in helping them.
-    - Make a one liner closing sentence to transition to the next subtitle.
-    -  Here's a list of the subtitle of this article to help you do logical transition between the content, 
-    they are in order, so the subtitle next to ${subtitle} in the list is the subtitle of the next paragraph of the article: ${listSubtitle}.
-    - Don't mention the target audience in the text, this is only to guide the narration.
-    `,
-    user: `
-    Create content for the subtitle "${subtitle}, your target audience for this is: "${targetAudience}, the mission is: "${mission}".
-    The context of the article is, use the context as instructions: "${context}".
-    `,
-  };
-};
-
-export const getIntroPrompt = async (
-  intro: string,
-  mission: string,
-  context: string,
-  targetAudience: string,
-  subtitle: string[],
-  lang: string,
-): Promise<{ system: string; user: string }> => {
-  return {
-    system: `Write introduction for the title: "${intro}" and the mission ${mission}", you're a professional on the subject.
-    Respect the requirement that follow: "
-
-    - Use bullet point to present de rest of the article.
-    - Use less then 5 bullet point.
-    - Use the list of subtitle to create logical transition in the intro and the rest of the article: ${subtitle}
-    - The introduction should be engaging and informative.
-    - The introduction must be designed in a way to keep the reader on the site.
-    - The content should be designed to hold the reader’s attention and provide valuable information without overwhelming them.
-    - The text output is destined to bring organic traffic to the website.
-    - It should have a lot of different keywords related to the context.
-    - if you have to explain a concept, make it clear and easy to understand.
-    - Don't mention the target audience in the text, this is only to guide the narration.
-    ${lang === 'French' ? '- Only use uppercase for the first letter of the title. Only the first letter of the first word should be capitalized, for the other word:  first letter should be lowercased. for normal use.' : ''}
-    - Make sure the text in the language: ${lang}.
-
-    "
-    `,
-    user: `Create an introduction respecting the requirements for the introduction: "${intro},
-    your target audience for this is: "${targetAudience}, 
-    Keep in mind that the mission is ${mission} and the context is "${context}".
-    Introduction should be short and concise, it should introduce the reader to the article.
-    `,
-  };
-};
-export const getConclusionPrompt = async (
-  closure: string,
-  mission: string,
-  context: string,
-  targetAudience: string,
-  subtitle: string[],
-  lang: string,
-  CTA: string,
-  domain: string,
-): Promise<{ system: string; user: string }> => {
-  return {
-    system: `Create conclusion with all the information you will be given, you're a professional on this context.
-    You will receive the subtitle of the conclusion, the mission, the context and the target audience.
-    You will respect the requirement that follow:
-    "
-
-    - Don't mention the call to action in the title.
-    - It should have an interesting and engaging title.
-    - Make sure the conclusion is related to the mission.
-    - The conclusion title should be pertinent to the article.
-    - Make sure the conclusion title is interesting and engaging.
-    - The conclusion subtitle should begin with "Conclusion: ".
-    - You will also receive a list of subtitles of this article to help you create a logical conclusion with the rest of the article.
-    - The text output is destined to bring organic traffic to a website.
-    - Summarize the key points and encourage the reader to take action, about the CTA (call to action): "${CTA}".
-    - Use the call to action in a intuitive way.
-    - Make sure the conclusion is related to the mission of the article: "${mission}".
-    - Link the domain of the client for further information: "${domain}", use a link html tag to the website, don't mention the url in the text, use by example "click here for further information". 
-    - Make the link in the conclusion in a way that the reader will want to click on it. 
-    - Make it concise and informative.
-    - If you have to explain a concept, make it clear and easy to understand
-    - The call to action should be related to the mission.
-    - The Call to Action (CTA) must be in a different paragraph, at last.
-    - Don't use word for word the information given to you.
-    - Don't mention the target audience in the text, this is only to guide the narration.
-    - Put the conclusion as a title in the closure.
-    ${lang === 'French' ? '- Only use uppercase for the first letter of the title. Only the first letter of the first word should be capitalized, for the other word:  first letter should be lowercased. for normal use.' : ''}
-    - Make sure the text in the language: ${lang}
-    "
-    `,
-    user: `Create a conclusion respecting the requirements for the conclusion: "${closure},
-    your target audience is: "${targetAudience}:,
-    the mission is: "${mission}" and the context is "${context}".
-    Here's a list of the subtitle of this article to help you do logical transition between the content: ${subtitle}.
-    `,
-  };
-};
-
-export const improveConclusionPrompt = async (
-  closure: string,
-  mission: string,
-  prompt: string,
-  targetAudience: string,
-  subtitle: string[],
-  lang: string,
-  CTA: string,
-  domain: string,
-  conclusion: string,
-): Promise<{ system: string; user: string }> => {
-  return {
-    system: `Edit the given conclusion to make sure it respect these requirements: 
-    "
-    - The conclusion title should be pertinent to the article.
-    - Summarize the key points and encourage the reader to take action, about the CTA (call to action): "${CTA}".
-    - Use the call to action in a intuitive way.
-    - Make sure the conclusion is related to the mission of the article: "${mission}".
-    - Link the domain of the client for further information: "${domain}", use a link html tag to the website, don't mention the url in the text, use by example "click here for further information". 
-    - Make the link in the conclusion in a way that the reader will want to click on it. 
-    - The call to action should be related to the mission.
-    - The Call to Action (CTA) must be in a different paragraph, at last.
-    - Don't use word for word the information given to you.
-    - Don't mention the target audience in the text, this is only to guide the narration.
-    "
-    `,
-    user: `Create a conclusion respecting the requirements for the conclusion: "${closure},
-    your target audience is: "${targetAudience}:,
-    the mission is: "${mission}" and the context is "${prompt}".
-    Here's the conclusion you will improve: "${conclusion}".
-    `,
-  };
-};
-
-export const improveBodyPrompt = async (
-  content: string,
-  lang: string,
-  listOfSubtitles: string[],
-  context: string,
-): Promise<{ system: string; user: string }> => {
-  return {
-    system: `You will receive the body of an article, you will improve it with the requirements that follow:
-    - Make sure the text is not redundant, easy to read, without big paragraph.
-    - Each part of the body must have at least 4 paragraphs.
-    - Make logic transition each part of the body.
-    - Make sure the language of the text is in ${lang}, even the subtitle.
-    ${lang === 'French' ? '- Only use uppercase for the first letter of the title. Only the first letter of the first word should be capitalized, for the other word:  first letter should be lowercased. for normal use.' : ''}
-    - Develop and explain each ideas.
-    - Make sure that the information of the text is display is various way.
-    - Make sure bullet points are used in the text. Add them in a MAXIMUM of 2 or 3 SECTIONS.
-    - Make sure that the text looks like a professional article made by a human.
-    - Make sure that the different section doesn't have the same structure, make sure that the different section don't look alike.
-    - You will need to separate the text in small paragraphs.
-    - Check the length of each paragraph, if one one is too long, split it.
-    - The text output is destined to bring organic traffic to a website.
-    `,
-    user: `Here's the content of the body of the article: "${content}" and the multiple subtitles of the article: ${listOfSubtitles}.
-    Your main job is to make sure the text is not redundant.
-    Make sure the body respect the instruction of the context: "${context}".
-    `,
-  };
-};
-
 export const editContentPrompt = async (
   article: string,
   lang: string,
@@ -306,7 +127,7 @@ export const editContentPrompt = async (
     - If acronyms are used make sure the acronyms is written in capital letters.
     - Make sure to format it in html format.    
     - Here's a list of the subtitle of this article to help you do logical transition between the content: "${subtitleList}" and to help you format the text in html format.
-    - You will add the title in a h1 tag at the beginning of the article: "${greatestTitle}".
+    - You will add this string as title for the article in a h1 tag: "${greatestTitle}".
     - Make sure that all the text is in language: "${lang}", even the subtitle.
     - If the article use abbreviations, make sure to write the full word and the abbreviation in parenthesis.
     - THE FIRST <P></P> MUST BE A SHORT INTRODUCTION OF THE ARTICLE. THIS SHOULD BE A MAXIMUM OF 2 SENTENCES.
@@ -327,7 +148,8 @@ export const editContentPrompt = async (
     - Make sure to format it in html format.
     Highlight the important words in bold.
     Make sure the article is easy to read and non-redundant.
-    Make sure there is not bullet point in each section, if there is, reformulate the bullet to change the way it's displayed.`,
+    Make sure there is not bullet point in each section, if there is, reformulate the bullet to change the way it's displayed.
+    Make sure paragraph are short and concise and that there are multiple in each section of the article.`,
   };
 };
